rm unnecessary bootloader_env.rs
